+++ bitframe-authentication/services/core/src/commonMain/kotlin/bitframe/authentication/signin/SignInService.kt
@@ -6,5 +6,10 @@ import later.Later
 import kotlin.js.JsExport
 
 abstract class SignInService {
+    fun validate(credentials: SignInCredentials) {
+        require(credentials.alias.isNotEmpty()) { "loginId (i.e. email/phone/username), must not be empty" }
+        require(credentials.password.isNotEmpty()) { "Password must not be empty" }
+    }
+
     abstract fun signIn(credentials: SignInCredentials): Later<LoginConundrum>
 }