+++ bitframe-authentication/services/core/src/commonMain/kotlin/bitframe/authentication/signin/SignInServiceImpl.kt
@@ -1,9 +1,9 @@
 package bitframe.authentication.signin
 
-import bitframe.authentication.config.ServiceConfig
 import bitframe.authentication.spaces.SpacesDao
 import bitframe.authentication.users.UsersDao
 import bitframe.daos.conditions.contains
+import bitframe.service.config.ServiceConfig
 import later.Later
 import later.await
 import later.later
@@ -15,6 +15,7 @@ class SignInServiceImpl(
 ) : SignInService() {
     private val scope = config.scope
     override fun signIn(credentials: SignInCredentials): Later<LoginConundrum> = scope.later {
+        validate(credentials)
         val matches = usersDao.all(where = "contacts" contains credentials.alias).await()
         if (matches.isEmpty()) throw RuntimeException("User with loginId=${credentials.alias}, not found")
         val match = matches.first()