+++ bitframe-server/frameworks/core/src/commonMain/kotlin/bitframe/server/modules/authentication/AuthenticationServiceImpl.kt
@@ -1,6 +1,5 @@
 package bitframe.server.modules.authentication
 
-import bitframe.authentication.config.ServiceConfig
 import bitframe.authentication.signin.SignInService
 import bitframe.authentication.signin.SignInServiceImpl
 import bitframe.authentication.spaces.SpacesDao
@@ -8,12 +7,13 @@ import bitframe.authentication.users.UsersDao
 import bitframe.authentication.users.UsersService
 import bitframe.authentication.users.UsersServiceImpl
 import bitframe.server.data.DAOProvider
+import bitframe.service.config.ServiceConfig
 
 class AuthenticationServiceImpl(
     private val usersDao: UsersDao,
     private val spacesDao: SpacesDao,
     override val signIn: SignInService = SignInServiceImpl(usersDao, spacesDao, ServiceConfig()),
     override val users: UsersService = UsersServiceImpl(usersDao, spacesDao)
-) : AuthenticationService{
+) : AuthenticationService {
     constructor(provider: DAOProvider) : this(provider.users, provider.spaces)
 }