+++ buildSrc/src/main/kotlin/versions.kt
@@ -24,14 +24,13 @@ object vers {
     }
 
     object bitframe {
-        val current = "0.0.28"
-        val previous = "0.0.25"
+        val current = "0.0.29"
+        val previous = "0.0.28"
     }
 
     object asoft {
         val foundation = "1.4.0"
         val platform = foundation
-        val access = "0.0.20"
         val contacts = "0.0.30"
         val duality = "0.0.50"
         val collections = foundation