+++ pi-monitor/pi-monitor-client/sdks/core/src/commonMain/kotlin/pimonitor/evaulation/business/BusinessService.kt
@@ -1,6 +1,6 @@
 package pimonitor.evaulation.business
 
-import bitframe.MiniService
+import bitframe.service.MiniService
 import later.Later
 import pimonitor.Monitor