+++ pi-monitor/pi-monitor-client/sdks/core/src/jvmTest/kotlin/acceptance/utils/IntegrationTest.kt
@@ -1,6 +1,6 @@
 package acceptance.utils
 
-import bitframe.authentication.KtorClientConfiguration
+import bitframe.service.config.KtorClientConfiguration
 import testing.ContainerTest
 
 open class IntegrationTest : ContainerTest() {