+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.0.29
+
+## Bitframe Authentication Service
+
+- [[BF94]](https://github.com/picortex/bitframe/issues/94) Lifted authentication ktor service away from bitframe client
+
+## Bitframe Client Sdk
+
+### Bug Fixes
+
+- [[BF92]](https://github.com/picortex/bitframe/issues/92) Fixed sign in validation bug for empty sign in credentials
+
 # 0.0.28
 
 ## CI & CD